Remove "homepage" in package.json
In the production build, the ejected CRA webpack config grabs the URL from the "homepage" key of the package.json and assumes the path is the path of all files at the deployed app. Our "homepage" defaulted to the github repo (github.com/azavea/pwd-fishway) resulting in a basepath of azavea/pwd-fishway instead of just the root. Removing the "homepage" key from package.json sets a default basepath of "/".

Refs #2