Ensure fonts are only requested once
Due to a bug in Styled Components
(https://github.com/styled-components/styled-components/issues/1593),
custom fonts were requested each time a component was mounted that
referenced them. This resulted in the text on the site flickering when
the fonts reloaded unnecessarily each time a new tab was clicked.

One suggested workaround is to load the fonts through a CSS stylesheet,
outside of the Styled Components pipeline. That workaround is
implemented here.

Refs #63