Process images and videos through webpack
To reduce the number of resource requests, process the fish photos
through webpack. This results in most of the images being Base64
encoded. These requests are faster and are cached. For some reason, with
the current implementation, the images are requested each time the
component is mounted. This does not happen when the images are processed
through Webpack. The videos are also processed through Webpack for
consistency.

Refs #63