Add live feed feature of SeeTheFishway
Prepare the app to accomodate the AXIS camera public IP live feed of the Fishway. Said feed won't be available until the camera is installed tentatively for the 2020 fish migration season. This work was done now rather than in the future to ensure the app is prepared, and save on mental spin up time.

Refs #74