Ignore right click, long click, and pinch-zoom for touchscreen optimization
A website in Chrome is not optimized for a museum display nor touchscreen by default. Chrome, and Windows, come built in with click and swipe interactions that can cause undesirable behavior to a museum user. We are able to use code and browser settings to manage these issues rather than rely on a 3rd party service like kiosk to try to host the app natively or otherwise. Some of the addressed behaviors are informed by the current app that educators and the client report as broken. Some undesirable touchscreen + browser behaviors such as overscroll are set by Chrome itself and are not part of this PR.

Refs #17