Show live feed error message on error only after feature has been seen
If the app is offline and the user navigates to the SeeTheFishway tab, the live feed won't show at all. However, if the app is offline while the user is on the SeeTheFishway tab seeing the live feed option, we've shown our cards and removing the option from the UI would be a weird UX. The goal is to not have the user feel like they're missing out on features.

Refs #77