Recreate videoplayer component when video is swapped out
This is a bug. Smart react re-rendering is not our desired use-case here. If the rendered component receives new props, we want to reset the state, but React will leave them as-is. This leaves the controls in an mis-matching state to the new, auto-playing video. React will recreate the component fresh if the `key` changes.