Widen pool of incorrect choices during quiz
There is only video for a subset of river fish. That subset is capable of forming quiz question answers, but we can populate the incorrect choices with any fish. For video-less fish in Meet The Fish, you never see them again thoroughout the app so I thought it'd be nice to include them in the quiz where possible. This makes some questions easier, i.e. an incorrect option is the obviously wrong eel, but also potentially harder, i.e. two catfish options are possible.

Also, refactor the custom fish prop types.

Refs #114